refactor(application): split primitives production by port owners
`primitives/production.rs` (4k lines) carried every production primitive
port beside the shared page/evidence helpers and 1.2k lines of tests.
Move each owner, unchanged, into `primitives/production/`:

- `lexical_grep`: the lexical grep and redundancy authorities
- `test_primitive`: the managed test-run primitive port
- `retrieval_ports`: the source-line and operational-health ports
- `extended_primitive`: the extended primitive port and durable storage
  status history
- `symbol_graph_snapshot`: the symbol-graph cursor snapshot authority
- `affected_tests`: the affected-tests port and attribution evidence
- `unavailable_evidence_tests`, `storage_table_detail_tests`,
  `affected_tests_tests`: the former inline test modules

The root keeps the shared page/coverage/graph-read helpers, the runtime
open path, and re-exports every public name at its old path. Moved
root-private items become `pub(super)`, the scope they already had.
